Bloco3

+++ bomservico_be/src/main/java/unoeste/fipp/bomservico/repositories/InteresseRepository.java
@@ -1,7 +1,8 @@
 package unoeste.fipp.bomservico.repositories;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import unoeste.fipp.bomservico.entities.Interesse;
 
 @Repository
-public interface InteresseRepository extends JpaRepository<Interesse, Long> { }
+public interface InteresseRepository extends JpaRepository<Interesse, Integer> { }
